refactor(modules): optimize prefs module
